+++ CHANGELOG.md
@@ -10,6 +10,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 - JoinCurves method in JoinOperator
 - SplitSurface method in SplitOperator (U and V direction splitting)
+- Simple BMP Exporter
+- ElevateDegree for surfaces in DegreeOperator
+- ReduceDegree for surfaces in DegreeOperator
 
 ### Changed
 - Vector3Double, Vector4Double class -> readonly struct